test: improve integration tests file structure
