+++ test/integration/asserts_test.go
@@ -1,0 +1,61 @@
+//go:build integration_tests
+// +build integration_tests
+
+package integration
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	v1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+
+	operatorv1alpha1 "github.com/kong/gateway-operator/api/v1alpha1"
+	"github.com/kong/gateway-operator/internal/consts"
+	k8sutils "github.com/kong/gateway-operator/internal/utils/kubernetes"
+)
+
+// mustListDataPlaneDeployments is a helper function for tests that
+// conveniently lists all deployments managed by a given dataplane.
+func mustListDataPlaneDeployments(t *testing.T, dataplane *operatorv1alpha1.DataPlane) []v1.Deployment {
+	deployments, err := k8sutils.ListDeploymentsForOwner(
+		ctx,
+		mgrClient,
+		consts.GatewayOperatorControlledLabel,
+		consts.DataPlaneManagedLabelValue,
+		dataplane.Namespace,
+		dataplane.UID,
+	)
+	require.NoError(t, err)
+	return deployments
+}
+
+// mustListControlPlaneDeployments is a helper function for tests that
+// conveniently lists all deployments managed by a given controlplane.
+func mustListControlPlaneDeployments(t *testing.T, controlplane *operatorv1alpha1.ControlPlane) []v1.Deployment {
+	deployments, err := k8sutils.ListDeploymentsForOwner(
+		ctx,
+		mgrClient,
+		consts.GatewayOperatorControlledLabel,
+		consts.ControlPlaneManagedLabelValue,
+		controlplane.Namespace,
+		controlplane.UID,
+	)
+	require.NoError(t, err)
+	return deployments
+}
+
+// mustListServices is a helper function for tests that
+// conveniently lists all services managed by a given dataplane.
+func mustListDataPlaneServices(t *testing.T, dataplane *operatorv1alpha1.DataPlane) []corev1.Service {
+	services, err := k8sutils.ListServicesForOwner(
+		ctx,
+		mgrClient,
+		consts.GatewayOperatorControlledLabel,
+		consts.DataPlaneManagedLabelValue,
+		dataplane.Namespace,
+		dataplane.UID,
+	)
+	require.NoError(t, err)
+	return services
+}
